fix: split play page into server wrapper + client component for useSearchParams
export const dynamic only works in Server Components — it is silently ignored
when exported from a "use client" file. Split the page so page.tsx is a Server
Component that owns the dynamic export and the Suspense boundary, and
PlayClient.tsx holds the interactive logic with useSearchParams().
